refactor: move authentication type definitions
