+++ packages/client/test/test-utils/fake/fakeEnvironment.ts
@@ -3,7 +3,7 @@ import { container, DependencyContainer } from 'tsyringe'
 import { BrubeckNode } from '../../../src/BrubeckNode'
 import { ConfigInjectionToken, createStrictConfig, StreamrClientConfig, StrictStreamrClientConfig } from '../../../src/Config'
 import { DestroySignal } from '../../../src/DestroySignal'
-import { AuthConfig } from '../../../src/Ethereum'
+import { AuthConfig } from '../../../src/Authentication'
 import { StorageNodeRegistry } from '../../../src/registry/StorageNodeRegistry'
 import { StreamrClient } from '../../../src/StreamrClient'
 import { initContainer } from '../../../src/Container'
